+++ src/uiobjects/thread.h
@@ -7,8 +7,8 @@
 #define THREAD_NOT_INITIATED 1
 #define THREAD_FINISHED 0
 
-#include "Constant.h"
-//#include "gmutex.h"
+#include "uiobjects/Constant.h"
+//#include "uiobjects/gmutex.h"
 #include <mutex>
 
 
@@ -23,8 +23,8 @@ protected:
         pthread_t hThread;
         pthread_attr_t attr;
         void *threadReturn;
-        
-        
+
+
 private:
     uint32_t    threadID;     // thread id - 0 until started
     T*          object;          // the object which owns the method
@@ -36,7 +36,7 @@ private:
     std::mutex hSingleStart;
     std::mutex hSingleRun;
     uint32_t ret;
-    
+
 
     // This function gets executed by a concurrent thread.
     static void *run(void * thread_obj)
@@ -45,8 +45,8 @@ private:
 //        const std::lock_guard<std::mutex> lock(thread->hSingleRun);
         thread->ret = (thread->object->*thread->method) ();
         pthread_attr_destroy(&thread->attr);
-        pthread_exit((void*)thread->ret);
-        return (void*)thread->ret;
+        pthread_exit((void*) (uintptr_t) thread->ret);
+        return (void*) (uintptr_t) thread->ret;
     }
     // Prevent copying of threads: No sensible implementation!
     Thread(const Thread<T>& other) {}
@@ -68,7 +68,7 @@ public:
     ~Thread(void){
         //std::cout << "Thread::Destructor"<<std::endl;
         //pthread_mutex_destroy(&hSingleStart);
-        
+
     }
 // -----------------------------------------------------------------------------
     T* getObject(){
@@ -112,7 +112,7 @@ public:
 //        std::cout << "  exiting with threadReturn :" << threadReturn << std::endl;
         return *((int*)(&threadReturn));
     }
-    
+
 // -----------------------------------------------------------------------------
     inline bool isRunning(){
         return getStatus() == THREAD_STILL_ACTIVE;
@@ -124,7 +124,7 @@ public:
         return hThread;
     }
 // -----------------------------------------------------------------------------
-    inline uint32_t getThreadID(){   
+    inline uint32_t getThreadID(){
         pthread_t ptid = pthread_self();
         memcpy(&this->threadID, &ptid, std::min(sizeof(this->threadID), sizeof(ptid)));
         return this->threadID;
@@ -137,7 +137,7 @@ public:
     uint32_t inline getStatus(){
         return this->status;
     }
-    
+
     inline void setStatus(uint32_t var){
         this->status = var;
 //        std::cout << "status asignning: " << status << std::endl;
